+++ AeSys/EoViPrimitiveSubModeMend.cpp
@@ -36,7 +36,7 @@ void AeSysView::OnModePrimitiveMend() {
 		OdDbBlockTableRecordPtr BlockTableRecord = Database()->getModelSpaceId().safeOpenObject(OdDb::kForWrite);
 		m_PrimitiveToMendCopy = m_PrimitiveToMend->Clone(BlockTableRecord);
 		m_MendPrimitiveBegin = m_PrimitiveToMend->SelectAtControlPoint(this, ptView);
-		m_MendPrimitiveVertexIndex = 1 << EoDbPrimitive::ControlPointIndex();
+		m_MendPrimitiveVertexIndex = static_cast<unsigned long>(1 << EoDbPrimitive::ControlPointIndex());
 
 		theApp.LoadModeResources(ID_MODE_PRIMITIVE_MEND);
 	}
